Extract Merge Responses from QueryHandler, add a test for it, remove date formatting for UHT (needs to be added in frontend)
