+++ api/lib/backends/UHT-Contacts/FetchRecord.js
@@ -55,10 +55,8 @@ const processCustomer = results => {
       if (result.tag_ref) {
         return {
           tagRef: checkString(result.tag_ref),
-          startDate: result.start_date
-            ? formatRecordDate(result.start_date)
-            : null,
-          endDate: result.end_date ? formatRecordDate(result.end_date) : null,
+          startDate: result.start_date,
+          endDate: result.end_date,
           tenure: checkString(result.tenure),
           currentBalance: result.current_balance,
           rentAmount: result.rent,
@@ -80,7 +78,7 @@ const processCustomer = results => {
     .filter(x => x);
 
   customer.tenancies = tenancies.map(tenancy => {
-    if (tenancy.endDate.includes('1900-01-01')) {
+    if (tenancy.endDate.getYear === '1990') {
       tenancy.endDate = null;
     }
     return tenancy;
